refactor: preserve existing deployment contracts
