+++ apps/backend/.env.development
@@ -70,7 +70,7 @@ HEXCLAVE_SVIX_API_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2NTUxNDA2M
 # Trusted reverse proxy for reading real client IP addresses.
 # Set to "vercel", "cloudflare", "cloudrun", or "generic". Generic proxies must
 # block direct origin access and overwrite X-Real-IP, X-Forwarded-Host, and X-Forwarded-Proto.
-STACK_TRUSTED_PROXY=
+HEXCLAVE_TRUSTED_PROXY=
 
 HEXCLAVE_ARTIFICIAL_DEVELOPMENT_DELAY_MS=500
