+++ apps/backend/package.json
@@ -20,7 +20,7 @@
     "build": "pnpm run codegen && tsdown --config tsdown.config.ts",
     "docker-build": "pnpm run codegen && tsdown --config tsdown.config.ts",
     "build-self-host-migration-script": "tsdown --config scripts/db-migrations.tsdown.config.ts",
-    "start": "pnpm run with-env:prod node dist/server.mjs",
+    "start": "node dist/server.mjs",
     "codegen-prisma": "HEXCLAVE_DATABASE_CONNECTION_STRING=\"${HEXCLAVE_DATABASE_CONNECTION_STRING:-${STACK_DATABASE_CONNECTION_STRING:-placeholder-database-connection-string}}\" pnpm run prisma generate",
     "codegen-prisma:watch": "HEXCLAVE_DATABASE_CONNECTION_STRING=\"${HEXCLAVE_DATABASE_CONNECTION_STRING:-${STACK_DATABASE_CONNECTION_STRING:-placeholder-database-connection-string}}\" pnpm run prisma generate --watch",
     "generate-private-sign-up-risk-engine": "pnpm run with-env tsx scripts/generate-private-sign-up-risk-engine.ts",
