+++ apps/backend/src/app/health/route.tsx
@@ -1,20 +1,13 @@
 import { globalPrismaClient } from "@/prisma-client";
+import { HexclaveAssertionError } from "@hexclave/shared/dist/utils/errors";
 
 export async function GET(req: Request) {
-  return await getHealthResponse(req, checkDatabaseConnection);
-}
-
-async function checkDatabaseConnection(): Promise<void> {
-  await globalPrismaClient.$primary().project.findFirst({
-    select: { id: true },
-  });
-}
-
-async function getHealthResponse(req: Request, checkDatabase: () => Promise<unknown>) {
   if (new URL(req.url).searchParams.get("db")) {
-    // A readiness probe needs database connectivity, not application data. An
-    // empty but correctly migrated database is ready and should pass this check.
-    await checkDatabase();
+    const project = await globalPrismaClient.project.findFirst({});
+
+    if (!project) {
+      throw new HexclaveAssertionError("No project found");
+    }
   }
 
   return Response.json({
@@ -28,27 +21,3 @@ async function getHealthResponse(req: Request, checkDatabase: () => Promise<unkn
     }
   });
 }
-
-const vitest = import.meta.vitest;
-vitest?.test("the database readiness probe verifies connectivity without requiring rows", async ({ expect }) => {
-  const { vi } = vitest;
-  const checkDatabase = vi.fn(async () => {});
-  const livenessResponse = await getHealthResponse(new Request("http://localhost/health"), checkDatabase);
-  const readinessResponse = await getHealthResponse(new Request("http://localhost/health?db=1"), checkDatabase);
-
-  expect({
-    livenessStatus: livenessResponse.status,
-    readinessStatus: readinessResponse.status,
-    databaseChecks: checkDatabase.mock.calls.length,
-  }).toEqual({
-    livenessStatus: 200,
-    readinessStatus: 200,
-    databaseChecks: 1,
-  });
-  await expect(getHealthResponse(
-    new Request("http://localhost/health?db=1"),
-    async () => {
-      throw new Error("database unavailable");
-    },
-  )).rejects.toThrowErrorMatchingInlineSnapshot(`[Error: database unavailable]`);
-});
